handle older subversion 1.6 versions
added code to handle subversion 1.6 where they put a .svn directory in
each of the directories in the repository. after removing all data from
the subversion repository and copying files from the git repository, we
need to rebuild the .svn directories in the previously established directories.

added a verbose flag.

added ability to specify a subversion repository name instead of a full url.

set the default location for temporary files to the caller's home
directory.